pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "changelog-feed"
version = "0.1.0"
description = "AI-driven CSA signal engine for GitHub, VS Code, and Visual Studio changelogs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "responses>=0.25.0",
]

[project.scripts]
changelog-feed = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
